Fix revert not returning (#105)
* Add revert reason

* Format code

* Fix typo

* Fix pr comment

* Format code

* Refactor in near and far call

* Add in near call to out of gas revert
