+++ src/op_handlers/far_call.rs
@@ -288,3 +288,13 @@ pub(crate) fn get_far_call_arguments(abi: U256) -> FarCallABI {
         is_system_call: system_call_byte != 0,
     }
 }
+
+pub fn perform_return(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+    let register = vm.get_register(opcode.src0_index);
+    let result = get_forward_memory_pointer(register.value, vm, register.is_pointer)?;
+    vm.set_register(
+        opcode.src0_index,
+        TaggedValue::new_pointer(FatPointer::encode(&result)),
+    );
+    Ok(())
+}
