+++ src/op_handlers/revert.rs
@@ -7,37 +7,36 @@ use crate::{
     Opcode,
 };
 
-use super::far_call::get_forward_memory_pointer;
+use super::far_call::{get_forward_memory_pointer, perform_return};
 
 pub fn revert(vm: &mut VMState, opcode: &Opcode) -> Result<bool, EraVmError> {
     vm.flag_eq = false;
     vm.flag_lt_of = false;
     vm.flag_gt = false;
-    if vm.running_contexts.len() > 1 || !vm.current_context()?.near_call_frames.is_empty() {
-        if !vm.current_context()?.near_call_frames.is_empty() {
-            // Near call
-            let previous_frame = vm.pop_frame()?;
-            if opcode.alters_vm_flags {
-                // Marks if it has .to_label
-                let to_label = opcode.imm0;
-                vm.current_frame_mut()?.pc = (to_label - 1) as u64;
-            // To account for the +1 later
-            } else {
-                vm.current_frame_mut()?.pc = previous_frame.exception_handler - 1;
-                // To account for the +1 later
-            }
-            vm.current_frame_mut()?.gas_left += previous_frame.gas_left;
+    if vm.in_near_call()? {
+        let previous_frame = vm.pop_frame()?;
+        if opcode.alters_vm_flags {
+            // Marks if it has .to_label
+            let to_label = opcode.imm0;
+            vm.current_frame_mut()?.pc = (to_label - 1) as u64;
+        // To account for the +1 later
         } else {
-            let register = vm.get_register(opcode.src0_index);
-            let result = get_forward_memory_pointer(register.value, vm, register.is_pointer)?;
-            vm.clear_registers();
-            vm.set_register(1, TaggedValue::new_pointer(FatPointer::encode(&result)));
-            vm.flag_lt_of = true;
-            let previous_frame = vm.pop_frame()?;
-            vm.current_frame_mut()?.pc = previous_frame.exception_handler;
+            vm.current_frame_mut()?.pc = previous_frame.exception_handler - 1;
+            // To account for the +1 later
         }
+        vm.current_frame_mut()?.gas_left += previous_frame.gas_left;
+        Ok(false)
+    } else if vm.in_far_call() {
+        let register = vm.get_register(opcode.src0_index);
+        let result = get_forward_memory_pointer(register.value, vm, register.is_pointer)?;
+        vm.clear_registers();
+        vm.set_register(1, TaggedValue::new_pointer(FatPointer::encode(&result)));
+        vm.flag_lt_of = true;
+        let previous_frame = vm.pop_frame()?;
+        vm.current_frame_mut()?.pc = previous_frame.exception_handler;
         Ok(false)
     } else {
+        perform_return(vm, opcode)?;
         Ok(true)
     }
 }
@@ -68,8 +67,7 @@ pub fn revert_out_of_gas(vm: &mut VMState) -> Result<(), EraVmError> {
     vm.flag_eq = false;
     vm.flag_lt_of = false;
     vm.flag_gt = false;
-    if !vm.current_context()?.near_call_frames.is_empty() {
-        // Near call
+    if vm.in_near_call()? {
         let previous_frame = vm.pop_frame()?;
         vm.current_frame_mut()?.pc = previous_frame.exception_handler - 1; // To account for the +1 later
         vm.current_frame_mut()?.gas_left += previous_frame.gas_left;
@@ -83,9 +81,9 @@ pub fn handle_error(vm: &mut VMState, err: EraVmError) -> Result<(), EraVmError>
     vm.flag_eq = false;
     vm.flag_lt_of = false;
     vm.flag_gt = false;
-    if !vm.current_context()?.near_call_frames.is_empty() {
+    if vm.in_near_call()? {
         revert_near_call(vm)?;
-    } else if vm.running_contexts.len() > 1 {
+    } else if vm.in_far_call() {
         revert_far_call(vm)?;
     } else {
         // Main context
